[IS2-130]: Integracion de simulación a la app de Operaciones. - Se integró todas los archivos dependientes de simulacion a la app de Operaciones. - Se modificó la metáfora, de forma que sea homogéneo 'conversiones' -> 'simulacion de operacion', tanto en backend como frontend. - Se adaptó los test de simulación. - Se adaptó los seeds.
