+++ backend/apps/operaciones/serializers.py
@@ -109,3 +109,29 @@ class TarjetaSerializer(serializers.ModelSerializer):
         model = Tarjeta
         fields = '__all__'  
 
+# ======================== Serializers para vistas de simulación de operación ========================
+"""
+Serializadores para validar los datos de entrada en las simulaciones de operaciones.
+"""
+
+class SimulacionPrivadaSerializer(serializers.Serializer):
+    """
+    Serializer para usuarios autenticados.
+    Requiere un cliente_id además de divisas, monto y método.
+    """
+    cliente_id = serializers.UUIDField()
+    divisa_origen = serializers.IntegerField()
+    divisa_destino = serializers.IntegerField()
+    monto = serializers.DecimalField(max_digits=30, decimal_places=2)
+    metodo_id = serializers.IntegerField()
+
+
+class SimulacionPublicaSerializer(serializers.Serializer):
+    """
+    Serializer para usuarios invitados.
+    No requiere cliente_id, solo datos básicos de simulación de operación.
+    """
+    divisa_origen = serializers.IntegerField()
+    divisa_destino = serializers.IntegerField()
+    monto = serializers.DecimalField(max_digits=30, decimal_places=2)
+    metodo_id = serializers.IntegerField()
