+++ backend/globalexchange/urls.py
@@ -54,5 +54,4 @@ urlpatterns = [
     path("api/auth/", include("apps.autenticacion.urls")),
     path('docs/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     path('redocs/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
-    path("api/conversiones/", include("apps.conversiones.urls")),
 ]
