+++ frontend/src/components/SimulacionOperacionPublica.tsx
@@ -1,17 +1,17 @@
   import { useState, useEffect } from "react";
   import {
-    simularConversionPublica,
+    simularOperacionPublica,
     getMetodosDisponibles,
-  } from "../services/conversionService";
+  } from "../services/simulacionService";
   import {
     type SimulacionResponse,
-  } from "../types/Conversion";
+  } from "../types/Simulacion";
   import { type Divisa } from "../types/Divisa";
   import { type MetodoFinanciero } from "../types/MetodoFinanciero";
   import { getDivisasConTasa } from "../services/divisaService";
   import { toast } from "react-toastify";
 
-  export default function SimulacionPublica() {
+  export default function SimulacionOperacionPublica() {
     const [monto, setMonto] = useState<number>(0);
     const [resultado, setResultado] = useState<SimulacionResponse | null>(null);
 
@@ -75,7 +75,7 @@
         return;
       }
       try {
-        const res = await simularConversionPublica({
+        const res = await simularOperacionPublica({
           divisa_origen: Number(divisaOrigen),
           divisa_destino: Number(divisaDestino),
           monto,
@@ -92,7 +92,7 @@
         <div className="w-full max-w-md bg-white rounded-xl shadow overflow-hidden border border-gray-200">
           {/* Encabezado */}
           <div className="bg-zinc-900 text-white text-center py-3 rounded-t-xl">
-            <h2 className="text-lg font-semibold">Simulación de Conversión</h2>
+            <h2 className="text-lg font-semibold">Simulación de Operación</h2>
           </div>
 
           <div className="p-6 space-y-5">
@@ -265,7 +265,7 @@
                   Operación: {resultado.operacion_cliente.toUpperCase()}
                 </div>
 
-                {/* Conversión */}
+                {/* Operación */}
                 <div className="text-center text-lg font-bold text-gray-900">
                   {resultado.monto_origen.toLocaleString()}{" "}
                   {resultado.divisa_origen} →{" "}
